backend/medexplain/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
pydantic==2.12.5
pydantic-settings==2.13.1
mistralai==2.3.1
httpx==0.28.1
python-dotenv==1.2.2
ffmpeg-python==0.2.0
elevenlabs==2.42.0
Pillow>=10.1.0
pytest==9.0.3
pytest-asyncio==1.3.0
python-multipart==0.0.18
pypdf==4.3.1
